test IdentifyRateLimiter in isolation
